+++ Source/Fuse.Maps/Android/FuseMap.java
@@ -17,8 +17,26 @@ import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.maps.model.Polyline;
+import com.google.android.gms.maps.model.PolylineOptions;
+import com.google.android.gms.maps.model.Polygon;
+import com.google.android.gms.maps.model.Circle;
+import com.google.android.gms.maps.model.CircleOptions;
+import com.google.android.gms.maps.model.PolygonOptions;
+import com.google.android.gms.maps.model.Cap;
+import com.google.android.gms.maps.model.ButtCap;
+import com.google.android.gms.maps.model.RoundCap;
+import com.google.android.gms.maps.model.SquareCap;
+import com.google.android.gms.maps.model.JointType;
+import com.google.android.gms.maps.model.Dash;
+import com.google.android.gms.maps.model.Gap;
+import com.google.android.gms.maps.model.Dot;
+import com.google.android.gms.maps.model.PatternItem;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import java.util.Map;
+import java.util.List;
+import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.HashMap;
 
 public class FuseMap extends FrameLayout {
@@ -42,6 +60,9 @@ public class FuseMap extends FrameLayout {
 	private MapView _mapView;
 	private boolean _isAnimating;
 	private Map<Marker, Integer> _markerIDs;
+	private List<Polyline> _polylines;
+	private List<Polygon> _polygons;
+	private List<Circle> _circles;
 
 	public FuseMap()
 	{
@@ -50,6 +71,9 @@ public class FuseMap extends FrameLayout {
 		_mapView = new MapView(com.fuse.Activity.getRootActivity());
 		addView(_mapView);
 		_markerIDs = new HashMap<Marker,Integer>();
+		_polylines = new ArrayList<Polyline>();
+		_polygons = new ArrayList<Polygon>();
+		_circles = new ArrayList<Circle>();
 
 		_mapView.getMapAsync(new OnMapReadyCallback()
 		{
@@ -60,12 +84,12 @@ public class FuseMap extends FrameLayout {
 			}
 		});
 	}
-	
+
 	public int getIdforMarker(Marker m)
 	{
 		return _markerIDs.get(m);
 	}
-	
+
 	@Override
 	public boolean dispatchTouchEvent(MotionEvent event) {
 		onTouch(event);
@@ -186,6 +210,129 @@ public class FuseMap extends FrameLayout {
 		return m.getId();
 	}
 
+	private Cap convertIntCap(int cap)
+	{
+		Cap capObj = null;
+		switch (cap)
+		{
+			case 0: capObj = new RoundCap(); break;
+			case 1: capObj = new ButtCap(); break;
+			case 2: capObj = new SquareCap(); break;
+			default:
+				capObj = new RoundCap();
+		}
+		return capObj;
+	}
+
+	private int selectJointType(int joinType)
+	{
+		int jointType;
+		switch (joinType) {
+			case 0: jointType = JointType.ROUND; break;
+			case 1: jointType = JointType.BEVEL; break;
+			case 2: jointType = JointType.DEFAULT; break;
+			default:
+				jointType = JointType.ROUND;
+		}
+		return jointType;
+	}
+
+	private Polyline drawPolyline(List<LatLng> points, int strokeColor, int lineWidth, boolean geodesic, int startCap, int endCap, int jointType, List<PatternItem> pattern)
+	{
+		Cap sCap = convertIntCap(startCap);
+		Cap eCap = convertIntCap(endCap);
+		PolylineOptions options = new PolylineOptions()
+			.color(strokeColor)
+			.width(lineWidth)
+			.startCap(sCap)
+			.endCap(eCap)
+			.jointType(jointType)
+			.geodesic(geodesic)
+			.zIndex(1)
+			.addAll(points);
+		if (pattern != null)
+			options.pattern(pattern);
+		return _googleMap.addPolyline(options);
+	}
+
+	private Polygon drawPolygon(List<LatLng> points, int strokeColor, int fillColor, int lineWidth, boolean geodesic, int jointType, List<PatternItem> pattern)
+	{
+		PolygonOptions options = new PolygonOptions()
+			.strokeColor(strokeColor)
+			.fillColor(fillColor)
+			.strokeWidth(lineWidth)
+			.strokeJointType(jointType)
+			.geodesic(geodesic)
+			.strokePattern(pattern)
+			.zIndex(1)
+			.addAll(points);
+		if (pattern != null)
+			options.strokePattern(pattern);
+		return _googleMap.addPolygon(options);
+	}
+
+	private Circle drawCircle(LatLng center, double radius, int strokeColor, int fillColor, int lineWidth, List<PatternItem> pattern)
+	{
+		CircleOptions options = new CircleOptions()
+			.strokeColor(strokeColor)
+			.fillColor(fillColor)
+			.strokeWidth(lineWidth)
+			.center(center)
+			.radius(radius)
+			.strokePattern(pattern)
+			.zIndex(1);
+		return _googleMap.addCircle(options);
+	}
+
+	private List<PatternItem> constructPattern(int[] dashPattern)
+	{
+		if (dashPattern.length == 2) {
+			if (dashPattern[0] > 0 && dashPattern[1] > 0)
+				return Arrays.asList(new Gap(dashPattern[1]), new Dash(dashPattern[0]));
+		}
+		return null;
+	}
+
+	public String AddOverlay(int type, double[] coordinates, int strokeColor, int fillColor, int lineWidth, boolean geodesic, int startCap, int endCap, int joinType, int[] dashPattern, double centerLatitude, double centerLongitude, double radius)
+	{
+		int jointType = selectJointType(joinType);
+		List<PatternItem> pattern = constructPattern(dashPattern);
+		List<LatLng> points = new ArrayList<>();
+		for (int i=0; i<coordinates.length; i+=2)
+			points.add(new LatLng(coordinates[i], coordinates[i+1]));
+		switch (type)
+		{
+			case 1:
+				Polygon polygon = drawPolygon(points, strokeColor, fillColor, lineWidth, geodesic, jointType, pattern);
+				_polygons.add(polygon);
+				return polygon.getId();
+			case 2:
+				Circle circle = drawCircle(new LatLng(centerLatitude, centerLongitude), radius, strokeColor, fillColor, lineWidth, pattern);
+				_circles.add(circle);
+				return circle.getId();
+			default:
+				Polyline polyline = drawPolyline(points, strokeColor, lineWidth, geodesic, startCap, endCap, jointType, pattern);
+				_polylines.add(polyline);
+				return polyline.getId();
+		}
+	}
+
+	public void clearOverlays()
+	{
+		for (Polyline polyline : _polylines) {
+			polyline.remove();
+		}
+		_polylines.clear();
+		for (Polygon polygon : _polygons) {
+			polygon.remove();
+		}
+		_polygons.clear();
+		for (Circle circle : _circles) {
+			circle.remove();
+		}
+		_circles.clear();
+	}
+
 	/* Camera */
 
 	public boolean isAnimating() { return _isAnimating; }
@@ -243,8 +390,9 @@ public class FuseMap extends FrameLayout {
 	}
 
 	public void clear(){
+		 for (Marker marker : _markerIDs.keySet())
+			marker.remove();
 		_markerIDs.clear();
-		_googleMap.clear();
 	}
 
 	private void zoomBy(float increment, double duration)